Replaced data storage and parsing with Pydantic to get automatic input file validation
